+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## Version 0.3.8
 - Add `Due_Date` to Task Structure
 - App now asks for `Due_Date` on task creation and task update
-# TODO
 - Show `days left` on board and color board background for task if it is overdue
 - Show `days left` of most urgent task for each board when changing boards
+- Add `freezegun` as test requirement
 
 ## Version 0.3.7
 - Bug Fix, `Complete_Time`-Key was wrong in Task Creation, when task was created with `kanban scan`